Add extremely cool way to dynamically change the ddc lsp timeout depending on how big the python file is
On large python files in the code base, lsp as a completion source
really struggles. 5 seconds seems to be the sweet spot for those large
files. Anything less, and the timeout will occur and no lsp candidates
will enter the completion menu. The trade-off is a longer time for the
popup to appear, so on small files, the timeout can be much smaller.

The file length itself isn't the determining factor so much as the
workspace, but this seems to help a bit.